pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "eml2png"
version = "1.0.0"
description = "Phishing Email Forensic Analyzer — generates cyber-infographic PNGs and interactive HTML reports"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4",
    "requests",
    "python-whois",
    "playwright",
]

[project.scripts]
eml2png = "eml2png.cli:main"
